requirements.txt
# Streamlit UI
streamlit>=1.32.0

# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# OpenAI for LLM and Whisper
openai>=1.50.0

# Speech Recognition (Whisper)
openai-whisper>=20231117

# Text-to-Speech
edge-tts>=6.1.0

# Audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
pydub>=0.25.1
numpy>=1.24.0

# Web interaction
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
httpx>=0.27.0

# Database (SQLite is built-in to Python, but adding useful extensions)
aiosqlite>=0.19.0

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Async support
asyncio>=3.4.3
